refactor(frontend): split 6.9k-line editor_preview.rs into submodules
Move four contiguous, self-contained leaf clusters out of editor_preview.rs
(6,946 -> 4,651 lines) into an editor_preview/ submodule directory, plus the
test module. The dense core and all shared types (PreviewScratch, consts,
FaceShade, FaceOutlineStyle, ...) stay in the module root.

  editor_preview/primitives.rs  projected-triangle submission: cull/clamp/split/OT insert
  editor_preview/backdrop.rs    frame clear, sky cyclorama, far-vista backdrop
  editor_preview/portals.rs     portal-seam outline rendering
  editor_preview/particles.rs   particle-emitter preview submission
  editor_preview/tests.rs       the module's unit tests

Child modules reach the root via `use super::*;`; moved items are `pub(crate)`.
No logic changes; `cargo check -p frontend` is clean.
