+++ emu/crates/frontend/src/editor_preview.rs
@@ -47,6 +47,16 @@ use psxed_project::{
 use crate::editor_textures::{EditorTextures, MaterialSlot};
 use psxed_ui::{PaintCellPreviewKind, ViewportCameraState};
 
+mod backdrop;
+mod particles;
+mod portals;
+mod primitives;
+
+use backdrop::*;
+use particles::*;
+use portals::*;
+use primitives::*;
+
 /// Maximum sectors we'll attempt to render in one preview pass.
 /// 64×64 grid would already be enormous for PSX (~16 MiB cooked); a
 /// 4096-cap caps the per-frame primitive count at a comfortable
@@ -2194,220 +2204,6 @@ fn walk_entities(
     }
 }
 
-fn push_particle_emitter_preview(
-    settings: &psxed_project::ParticleEmitterSettings,
-    origin: psx_engine::WorldVertex,
-    particle_slot: MaterialSlot,
-    preview_tick: u32,
-    scratch: &mut PreviewScratch,
-) {
-    if !settings.enabled
-        || settings.max_particles == 0
-        || settings.lifetime_frames == 0
-        || settings.spawn_rate_q8 == 0
-    {
-        return;
-    }
-    let lifetime = settings.lifetime_frames as u32;
-    let steady_count = ((settings.spawn_rate_q8 as u32)
-        .saturating_mul(lifetime)
-        .saturating_add(60 * 256 - 1))
-        / (60 * 256);
-    let count = (settings.max_particles as u32)
-        .min(PREVIEW_PARTICLE_DRAW_CAP as u32)
-        .min(steady_count.max(1));
-    if count == 0 {
-        return;
-    }
-    let mut i = 0u32;
-    while i < count {
-        let seed = preview_particle_seed(origin.x as u32, origin.z as u32, i);
-        let age = (preview_tick + (i * lifetime / count)) % lifetime;
-        push_particle_preview_sample(
-            settings,
-            origin,
-            particle_slot,
-            seed,
-            age as i32,
-            lifetime as i32,
-            scratch,
-        );
-        i += 1;
-    }
-}
-
-fn push_particle_preview_sample(
-    settings: &psxed_project::ParticleEmitterSettings,
-    origin: psx_engine::WorldVertex,
-    particle_slot: MaterialSlot,
-    seed: u32,
-    age: i32,
-    lifetime: i32,
-    scratch: &mut PreviewScratch,
-) {
-    let spawn_radius = settings.spawn_radius as i32;
-    let origin_x = origin
-        .x
-        .saturating_add(preview_particle_signed_spread(seed, spawn_radius));
-    let origin_y = origin.y.saturating_add(preview_particle_signed_spread(
-        seed.rotate_left(9),
-        spawn_radius / 2,
-    ));
-    let origin_z = origin.z.saturating_add(preview_particle_signed_spread(
-        seed.rotate_left(17),
-        spawn_radius,
-    ));
-    let x = preview_particle_axis_position(
-        origin_x,
-        settings.base_velocity_q4[0],
-        settings.random_velocity_q4[0],
-        settings.acceleration_q4[0],
-        age,
-        seed.rotate_left(3),
-    );
-    let y = preview_particle_axis_position(
-        origin_y,
-        settings.base_velocity_q4[1],
-        settings.random_velocity_q4[1],
-        settings.acceleration_q4[1],
-        age,
-        seed.rotate_left(11),
-    );
-    let z = preview_particle_axis_position(
-        origin_z,
-        settings.base_velocity_q4[2],
-        settings.random_velocity_q4[2],
-        settings.acceleration_q4[2],
-        age,
-        seed.rotate_left(21),
-    );
-    let projected = gte_scene::project_vertex(world_to_view([x, y, z]));
-    if projected.sz == 0 {
-        return;
-    }
-    let t_q8 = if lifetime <= 1 {
-        255
-    } else {
-        ((age * 255) / (lifetime - 1)).clamp(0, 255)
-    };
-    let size = preview_particle_lerp_u16(settings.start_size, settings.end_size, t_q8);
-    let half = ((i32::from(size) * PROJ_H) / i32::from(projected.sz.max(1))).clamp(
-        i32::from(PREVIEW_PARTICLE_MIN_SCREEN_SIZE),
-        i32::from(PREVIEW_PARTICLE_MAX_SCREEN_SIZE),
-    ) as i16;
-    let tint = preview_particle_lerp_rgb(settings.start_color, settings.end_color, t_q8);
-    push_particle_preview_quad(
-        projected,
-        half,
-        particle_slot,
-        tint,
-        psx_blend_mode(settings.blend_mode),
-        scratch,
-    );
-}
-
-fn push_particle_preview_quad(
-    center: psx_gte::scene::Projected,
-    half: i16,
-    particle_slot: MaterialSlot,
-    tint: (u8, u8, u8),
-    blend_mode: BlendMode,
-    scratch: &mut PreviewScratch,
-) {
-    if scratch.particle_used >= PREVIEW_PARTICLE_DRAW_CAP {
-        return;
-    }
-    let left = (i32::from(center.sx).saturating_sub(i32::from(half)))
-        .clamp(i16::MIN as i32, i16::MAX as i32) as i16;
-    let right = (i32::from(center.sx).saturating_add(i32::from(half)))
-        .clamp(i16::MIN as i32, i16::MAX as i32) as i16;
-    let top = (i32::from(center.sy).saturating_sub(i32::from(half)))
-        .clamp(i16::MIN as i32, i16::MAX as i32) as i16;
-    let bottom = (i32::from(center.sy).saturating_add(i32::from(half)))
-        .clamp(i16::MIN as i32, i16::MAX as i32) as i16;
-    if left == right || top == bottom {
-        return;
-    }
-    let u0 = PREVIEW_PARTICLE_TEXEL_U;
-    let v0 = PREVIEW_PARTICLE_TEXEL_V;
-    let u1 = u0 + particle_slot.texture_width.saturating_sub(1);
-    let v1 = v0 + particle_slot.texture_height.saturating_sub(1);
-    let material = preview_texture_material(particle_slot, tint, blend_mode);
-    let idx = scratch.particle_used;
-    scratch.particle_quads[idx] = QuadTexturedMaterial::with_material(
-        [(left, top), (right, top), (left, bottom), (right, bottom)],
-        [(u0, v0), (u1, v0), (u0, v1), (u1, v1)],
-        material,
-    );
-    scratch.particle_used = idx + 1;
-    let packet_ptr: *mut QuadTexturedMaterial = &mut scratch.particle_quads[idx];
-    unsafe {
-        scratch.ot.insert(
-            room_depth_slot(center.sz as u32),
-            packet_ptr.cast::<u32>(),
-            QuadTexturedMaterial::WORDS,
-        );
-    }
-}
-
-fn preview_particle_axis_position(
-    origin: i32,
-    base_velocity_q4: i16,
-    random_velocity_q4: u16,
-    acceleration_q4: i16,
-    age: i32,
-    seed: u32,
-) -> i32 {
-    let random_velocity = preview_particle_signed_spread(seed, random_velocity_q4 as i32);
-    let velocity = i32::from(base_velocity_q4).saturating_add(random_velocity);
-    let velocity_term = velocity.saturating_mul(age) >> 4;
-    let acceleration_term = i32::from(acceleration_q4)
-        .saturating_mul(age)
-        .saturating_mul(age)
-        >> 5;
-    origin
-        .saturating_add(velocity_term)
-        .saturating_add(acceleration_term)
-}
-
-fn preview_particle_seed(x: u32, z: u32, index: u32) -> u32 {
-    let mut value = x
-        .rotate_left(7)
-        .wrapping_add(z.rotate_left(17))
-        .wrapping_add(index.wrapping_mul(0x85EB_CA6B));
-    value ^= value >> 16;
-    value = value.wrapping_mul(0x7FEB_352D);
-    value ^= value >> 15;
-    value = value.wrapping_mul(0x846C_A68B);
-    value ^ (value >> 16)
-}
-
-fn preview_particle_signed_spread(seed: u32, spread: i32) -> i32 {
-    if spread <= 0 {
-        return 0;
-    }
-    let span = spread.saturating_mul(2).saturating_add(1) as u32;
-    (seed % span) as i32 - spread
-}
-
-fn preview_particle_lerp_u16(a: u16, b: u16, t_q8: i32) -> u16 {
-    let inv = 255 - t_q8;
-    (((i32::from(a) * inv) + (i32::from(b) * t_q8)) / 255).clamp(0, u16::MAX as i32) as u16
-}
-
-fn preview_particle_lerp_rgb(a: [u8; 3], b: [u8; 3], t_q8: i32) -> (u8, u8, u8) {
-    (
-        preview_particle_lerp_u8(a[0], b[0], t_q8),
-        preview_particle_lerp_u8(a[1], b[1], t_q8),
-        preview_particle_lerp_u8(a[2], b[2], t_q8),
-    )
-}
-
-fn preview_particle_lerp_u8(a: u8, b: u8, t_q8: i32) -> u8 {
-    let inv = 255 - t_q8;
-    (((i32::from(a) * inv) + (i32::from(b) * t_q8)) / 255).clamp(0, 255) as u8
-}
-
 fn walk_image_props(
     project: &ProjectDocument,
     room_id: psxed_project::NodeId,
@@ -4360,389 +4156,6 @@ fn push_ghost_wall_outline(
     }
 }
 
-fn walk_portal_seams(
-    project: &ProjectDocument,
-    room_id: NodeId,
-    grid: &WorldGrid,
-    hidden_scene_nodes: &HashSet<NodeId>,
-    selected: NodeId,
-    hovered: Option<NodeId>,
-    scratch: &mut PreviewScratch,
-) {
-    let scene = project.active_scene();
-    for node in scene.nodes() {
-        if !matches!(node.kind, NodeKind::Portal { .. })
-            || scene_node_hidden(scene, hidden_scene_nodes, node.id)
-            || !is_descendant_of_room(scene, node.id, room_id)
-        {
-            continue;
-        }
-        let mut style = PORTAL_SEAM_STYLE;
-        if node.id == selected || hovered == Some(node.id) {
-            style.thickness_px = 4.0;
-        }
-        push_portal_seam_edges(grid, portal_seam_edges_for_node(grid, node), style, scratch);
-    }
-}
-
-fn push_portal_seam_edges(
-    grid: &WorldGrid,
-    edges: Vec<PortalEdge>,
-    style: FaceOutlineStyle,
-    scratch: &mut PreviewScratch,
-) {
-    for run in portal_seam_runs(edges) {
-        push_portal_seam_run(grid, run, style, scratch);
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-struct PortalSeamRun {
-    start: PortalEdge,
-    len: u16,
-}
-
-fn portal_seam_runs(mut edges: Vec<PortalEdge>) -> Vec<PortalSeamRun> {
-    edges.retain(|edge| matches!(edge.direction, GridDirection::North | GridDirection::East));
-    edges.sort_by_key(|edge| match edge.direction {
-        GridDirection::North => (0, edge.z, edge.x),
-        GridDirection::East => (1, edge.x, edge.z),
-        _ => (2, edge.z, edge.x),
-    });
-    let mut runs = Vec::new();
-    for edge in edges {
-        if let Some(run) = runs.last_mut() {
-            if portal_edge_continues_run(run, edge) {
-                run.len = run.len.saturating_add(1);
-                continue;
-            }
-        }
-        runs.push(PortalSeamRun {
-            start: edge,
-            len: 1,
-        });
-    }
-    runs
-}
-
-fn portal_edge_continues_run(run: &PortalSeamRun, edge: PortalEdge) -> bool {
-    if edge.direction != run.start.direction {
-        return false;
-    }
-    match edge.direction {
-        GridDirection::North => {
-            edge.z == run.start.z && run.start.x.checked_add(run.len) == Some(edge.x)
-        }
-        GridDirection::East => {
-            edge.x == run.start.x && run.start.z.checked_add(run.len) == Some(edge.z)
-        }
-        _ => false,
-    }
-}
-
-fn push_portal_seam_run(
-    grid: &WorldGrid,
-    run: PortalSeamRun,
-    style: FaceOutlineStyle,
-    scratch: &mut PreviewScratch,
-) {
-    if run.len == 0 {
-        return;
-    }
-    let mut first: Option<[spatial::RoomPoint; 4]> = None;
-    let mut last: Option<[spatial::RoomPoint; 4]> = None;
-    for offset in 0..run.len {
-        let edge = match run.start.direction {
-            GridDirection::North => {
-                let Some(x) = run.start.x.checked_add(offset) else {
-                    return;
-                };
-                PortalEdge { x, ..run.start }
-            }
-            GridDirection::East => {
-                let Some(z) = run.start.z.checked_add(offset) else {
-                    return;
-                };
-                PortalEdge { z, ..run.start }
-            }
-            _ => return,
-        };
-        let Some(corners) = portal_edge_wall_corners(grid, edge) else {
-            continue;
-        };
-        push_portal_segment(scratch, corners[0], corners[1], style);
-        push_portal_segment(scratch, corners[3], corners[2], style);
-        if first.is_none() {
-            first = Some(corners);
-        }
-        last = Some(corners);
-    }
-
-    let Some(first) = first else {
-        return;
-    };
-    let Some(last) = last else {
-        return;
-    };
-    match run.start.direction {
-        GridDirection::North => {
-            push_portal_segment(scratch, first[0], first[3], style);
-            push_portal_segment(scratch, last[1], last[2], style);
-        }
-        GridDirection::East => {
-            push_portal_segment(scratch, first[1], first[2], style);
-            push_portal_segment(scratch, last[0], last[3], style);
-        }
-        _ => {}
-    }
-}
-
-fn push_portal_edge_wall_outline(
-    grid: &WorldGrid,
-    wcx: i32,
-    wcz: i32,
-    dir: GridDirection,
-    style: FaceOutlineStyle,
-    scratch: &mut PreviewScratch,
-) {
-    let Some(corners) = portal_edge_wall_corners_for_world_cell(grid, wcx, wcz, dir) else {
-        return;
-    };
-    push_portal_segment(scratch, corners[0], corners[1], style);
-    push_portal_segment(scratch, corners[1], corners[2], style);
-    push_portal_segment(scratch, corners[2], corners[3], style);
-    push_portal_segment(scratch, corners[3], corners[0], style);
-}
-
-fn portal_edge_wall_corners(grid: &WorldGrid, edge: PortalEdge) -> Option<[spatial::RoomPoint; 4]> {
-    portal_edge_wall_corners_for_world_cell(
-        grid,
-        grid.origin[0] + edge.x as i32,
-        grid.origin[1] + edge.z as i32,
-        edge.direction,
-    )
-}
-
-fn portal_edge_wall_corners_for_world_cell(
-    grid: &WorldGrid,
-    wcx: i32,
-    wcz: i32,
-    dir: GridDirection,
-) -> Option<[spatial::RoomPoint; 4]> {
-    const BOTTOM_LIFT: i32 = 24;
-    let bounds = spatial::cell_bounds_from_world_cell(wcx, wcz, grid.sector_size);
-    let heights = portal_edge_height_span_for_world_cell(grid, wcx, wcz, dir);
-    let mut corners = spatial::editor_wall_outline_corners(bounds, dir, heights, 0)?;
-    corners[0][1] = corners[0][1].saturating_add(BOTTOM_LIFT);
-    corners[1][1] = corners[1][1].saturating_add(BOTTOM_LIFT);
-    Some(corners)
-}
-
-fn portal_edge_height_span_for_world_cell(
-    grid: &WorldGrid,
-    wcx: i32,
-    wcz: i32,
-    dir: GridDirection,
-) -> [i32; 4] {
-    let mut bottom: [Option<i32>; 2] = [None, None];
-    let mut top: [Option<i32>; 2] = [None, None];
-    let mut fallback_bottom: Option<i32> = None;
-    let mut fallback_top: Option<i32> = None;
-
-    if let Some(sector) = sector_for_world_cell(grid, wcx, wcz) {
-        sample_sector_edge_span(sector, dir, false, &mut bottom, &mut top);
-        sample_sector_vertical_bounds(sector, &mut fallback_bottom, &mut fallback_top);
-    }
-    if let Some((nwcx, nwcz, opposite)) = portal_neighbour_world_cell(wcx, wcz, dir) {
-        if let Some(sector) = sector_for_world_cell(grid, nwcx, nwcz) {
-            sample_sector_edge_span(sector, opposite, true, &mut bottom, &mut top);
-            sample_sector_vertical_bounds(sector, &mut fallback_bottom, &mut fallback_top);
-        }
-    }
-
-    let fallback_bottom = fallback_bottom.unwrap_or(0);
-    let bottom = [
-        bottom[0].unwrap_or(fallback_bottom).min(fallback_bottom),
-        bottom[1].unwrap_or(fallback_bottom).min(fallback_bottom),
-    ];
-    let fallback_top = fallback_top.unwrap_or_else(|| {
-        bottom[0]
-            .max(bottom[1])
-            .saturating_add(grid.sector_size.max(1))
-    });
-    let mut top = [
-        top[0].unwrap_or(fallback_top).max(fallback_top),
-        top[1].unwrap_or(fallback_top).max(fallback_top),
-    ];
-    for i in 0..2 {
-        if top[i] <= bottom[i] {
-            top[i] = bottom[i].saturating_add(grid.sector_size.max(1));
-        }
-    }
-    [bottom[0], bottom[1], top[1], top[0]]
-}
-
-fn sector_for_world_cell(grid: &WorldGrid, wcx: i32, wcz: i32) -> Option<&GridSector> {
-    let (sx, sz) = grid.world_cell_to_array(wcx, wcz)?;
-    grid.sector(sx, sz)
-}
-
-fn portal_neighbour_world_cell(
-    wcx: i32,
-    wcz: i32,
-    dir: GridDirection,
-) -> Option<(i32, i32, GridDirection)> {
-    let opposite = dir.opposite_cardinal()?;
-    match dir {
-        GridDirection::North => Some((wcx, wcz.checked_add(1)?, opposite)),
-        GridDirection::East => Some((wcx.checked_add(1)?, wcz, opposite)),
-        GridDirection::South => Some((wcx, wcz.checked_sub(1)?, opposite)),
-        GridDirection::West => Some((wcx.checked_sub(1)?, wcz, opposite)),
-        GridDirection::NorthWestSouthEast | GridDirection::NorthEastSouthWest => None,
-    }
-}
-
-fn sample_sector_edge_span(
-    sector: &GridSector,
-    dir: GridDirection,
-    reverse: bool,
-    bottom: &mut [Option<i32>; 2],
-    top: &mut [Option<i32>; 2],
-) {
-    if let Some(edge) = sector
-        .floor
-        .as_ref()
-        .and_then(|floor| horizontal_edge_heights_for_portal(floor.heights, dir))
-    {
-        include_min_edge(bottom, maybe_reverse_edge(edge, reverse));
-    }
-    if let Some(edge) = sector
-        .ceiling
-        .as_ref()
-        .and_then(|ceiling| horizontal_edge_heights_for_portal(ceiling.heights, dir))
-    {
-        include_max_edge(top, maybe_reverse_edge(edge, reverse));
-    }
-    for wall in sector.walls.get(dir) {
-        let wall_bottom = [
-            wall.heights[WallCorner::BL.idx()],
-            wall.heights[WallCorner::BR.idx()],
-        ];
-        let wall_top = [
-            wall.heights[WallCorner::TL.idx()],
-            wall.heights[WallCorner::TR.idx()],
-        ];
-        include_min_edge(bottom, maybe_reverse_edge(wall_bottom, reverse));
-        include_max_edge(top, maybe_reverse_edge(wall_top, reverse));
-    }
-}
-
-fn sample_sector_vertical_bounds(
-    sector: &GridSector,
-    bottom: &mut Option<i32>,
-    top: &mut Option<i32>,
-) {
-    if let Some(floor) = &sector.floor {
-        for height in floor.heights {
-            include_min_value(bottom, height);
-        }
-    }
-    if let Some(ceiling) = &sector.ceiling {
-        for height in ceiling.heights {
-            include_max_value(top, height);
-        }
-    }
-    for dir in GridDirection::ALL {
-        for wall in sector.walls.get(dir) {
-            for height in wall.heights {
-                include_min_value(bottom, height);
-                include_max_value(top, height);
-            }
-        }
-    }
-}
-
-fn horizontal_edge_heights_for_portal(heights: [i32; 4], dir: GridDirection) -> Option<[i32; 2]> {
-    match dir {
-        GridDirection::North => Some([heights[Corner::NW.idx()], heights[Corner::NE.idx()]]),
-        GridDirection::East => Some([heights[Corner::NE.idx()], heights[Corner::SE.idx()]]),
-        GridDirection::South => Some([heights[Corner::SE.idx()], heights[Corner::SW.idx()]]),
-        GridDirection::West => Some([heights[Corner::SW.idx()], heights[Corner::NW.idx()]]),
-        GridDirection::NorthWestSouthEast | GridDirection::NorthEastSouthWest => None,
-    }
-}
-
-fn maybe_reverse_edge(mut edge: [i32; 2], reverse: bool) -> [i32; 2] {
-    if reverse {
-        edge.swap(0, 1);
-    }
-    edge
-}
-
-fn include_min_edge(target: &mut [Option<i32>; 2], edge: [i32; 2]) {
-    for i in 0..2 {
-        include_min_value(&mut target[i], edge[i]);
-    }
-}
-
-fn include_max_edge(target: &mut [Option<i32>; 2], edge: [i32; 2]) {
-    for i in 0..2 {
-        include_max_value(&mut target[i], edge[i]);
-    }
-}
-
-fn include_min_value(target: &mut Option<i32>, value: i32) {
-    *target = Some(target.map_or(value, |current| current.min(value)));
-}
-
-fn include_max_value(target: &mut Option<i32>, value: i32) {
-    *target = Some(target.map_or(value, |current| current.max(value)));
-}
-
-fn push_portal_segment(
-    scratch: &mut PreviewScratch,
-    a: spatial::RoomPoint,
-    b: spatial::RoomPoint,
-    style: FaceOutlineStyle,
-) {
-    let pa = gte_scene::project_vertex(world_to_view(a));
-    let pb = gte_scene::project_vertex(world_to_view(b));
-    if pa.sz == 0 || pb.sz == 0 {
-        return;
-    }
-    push_screen_line(scratch, pa, pb, style);
-}
-
-fn canonical_portal_edge_for_array_cell(
-    sx: u16,
-    sz: u16,
-    dir: GridDirection,
-) -> Option<PortalEdge> {
-    match dir {
-        GridDirection::North => Some(PortalEdge {
-            x: sx,
-            z: sz,
-            direction: GridDirection::North,
-        }),
-        GridDirection::East => Some(PortalEdge {
-            x: sx,
-            z: sz,
-            direction: GridDirection::East,
-        }),
-        GridDirection::South => Some(PortalEdge {
-            x: sx,
-            z: sz.checked_sub(1)?,
-            direction: GridDirection::North,
-        }),
-        GridDirection::West => Some(PortalEdge {
-            x: sx.checked_sub(1)?,
-            z: sz,
-            direction: GridDirection::East,
-        }),
-        GridDirection::NorthWestSouthEast | GridDirection::NorthEastSouthWest => None,
-    }
-}
-
 /// Hover and Selected outline styling. RGB plus screen-space line
 /// thickness in pixels. Keep these light: they are editor affordances,
 /// not scene geometry, and thick strokes obscure PS1-scale surfaces.
@@ -5233,1714 +4646,6 @@ fn bright_overlay_color(rgb: (u8, u8, u8)) -> egui::Color32 {
     egui::Color32::from_rgb(lift(rgb.0), lift(rgb.1), lift(rgb.2))
 }
 
-/// Stamp a GP0(02h) fill-rectangle into `scratch.clear_packet` and
-/// link it into the back-most OT slot so it runs first when DMA
-/// walks the chain -- which is the same pattern PS1 software uses to
-/// "clear" the framebuffer at the start of every frame, since the
-/// HwRenderer (faithfully) preserves VRAM across frames the way real
-/// hardware does.
-fn push_clear(scratch: &mut PreviewScratch, color: [u8; 3]) {
-    // PSX VRAM coords; the editor's HwRenderer renders the same
-    // 320×240 sub-rect that the runtime frame-buffer would land in.
-    let color_word =
-        0x0200_0000_u32 | color[0] as u32 | ((color[1] as u32) << 8) | ((color[2] as u32) << 16);
-    let xy_word = 0u32; // top-left at (0, 0)
-    let wh_word = ((240u32) << 16) | 320u32; // pack_xy(320, 240)
-                                             // word[0] is rewritten by `OrderingTable::insert` with the
-                                             // chain tag -- leave it at 0 here.
-    scratch.clear_packet[1] = color_word;
-    scratch.clear_packet[2] = xy_word;
-    scratch.clear_packet[3] = wh_word;
-    let ptr: *mut u32 = scratch.clear_packet.as_mut_ptr();
-    unsafe {
-        scratch.ot.insert(PREVIEW_CLEAR_SLOT, ptr, 3);
-    }
-}
-
-fn push_cyclorama(
-    scratch: &mut PreviewScratch,
-    sky: psxed_project::ResolvedSkySettings,
-    camera: psx_engine::WorldCamera,
-) {
-    if !sky.enabled {
-        return;
-    }
-    for quad in psxed_project::generate_sky_cyclorama(sky).into_iter().rev() {
-        let Some(projected) = preview_project_cyclorama_quad(quad.direction_q12, camera) else {
-            continue;
-        };
-        if preview_cyclorama_quad_outside_screen(projected) {
-            continue;
-        }
-        push_sky_quad_corners(scratch, projected, quad.rgb);
-    }
-}
-
-fn preview_project_cyclorama_quad(
-    dirs: [[i16; 3]; 4],
-    camera: psx_engine::WorldCamera,
-) -> Option<[(i16, i16); 4]> {
-    Some([
-        preview_project_cyclorama_direction(dirs[0], camera)?,
-        preview_project_cyclorama_direction(dirs[1], camera)?,
-        preview_project_cyclorama_direction(dirs[2], camera)?,
-        preview_project_cyclorama_direction(dirs[3], camera)?,
-    ])
-}
-
-fn preview_project_cyclorama_direction(
-    dir: [i16; 3],
-    camera: psx_engine::WorldCamera,
-) -> Option<(i16, i16)> {
-    let x = i32::from(dir[0]);
-    let y = i32::from(dir[1]);
-    let z = i32::from(dir[2]);
-    let sin_yaw = camera.sin_yaw.raw();
-    let cos_yaw = camera.cos_yaw.raw();
-    let sin_pitch = camera.sin_pitch.raw();
-    let cos_pitch = camera.cos_pitch.raw();
-    let x1 = mul_q12_i32(x, cos_yaw) - mul_q12_i32(z, sin_yaw);
-    let z1 = -mul_q12_i32(x, sin_yaw) - mul_q12_i32(z, cos_yaw);
-    let y2 = mul_q12_i32(y, cos_pitch) - mul_q12_i32(z1, sin_pitch);
-    let z2 = mul_q12_i32(y, sin_pitch) + mul_q12_i32(z1, cos_pitch);
-    if z2 <= NEAR_Z {
-        return None;
-    }
-    let sx = SCREEN_CX + (x1 * PROJ_H) / z2;
-    let sy = SCREEN_CY - (y2 * PROJ_H) / z2;
-    Some((sx.clamp(-512, 831) as i16, sy.clamp(-256, 495) as i16))
-}
-
-fn preview_cyclorama_quad_outside_screen(points: [(i16, i16); 4]) -> bool {
-    let min_x = points.iter().map(|p| p.0).min().unwrap_or(0);
-    let max_x = points.iter().map(|p| p.0).max().unwrap_or(0);
-    let min_y = points.iter().map(|p| p.1).min().unwrap_or(0);
-    let max_y = points.iter().map(|p| p.1).max().unwrap_or(0);
-    max_x < 0 || min_x >= SCREEN_W as i16 || max_y < 0 || min_y >= SCREEN_H as i16
-}
-
-fn push_sky_quad_corners(
-    scratch: &mut PreviewScratch,
-    points: [(i16, i16); 4],
-    colors: [[u8; 3]; 4],
-) {
-    if scratch.sky_used >= scratch.sky_quads.len() {
-        return;
-    }
-    let quad = QuadGouraud::new(
-        points,
-        [
-            (colors[0][0], colors[0][1], colors[0][2]),
-            (colors[1][0], colors[1][1], colors[1][2]),
-            (colors[2][0], colors[2][1], colors[2][2]),
-            (colors[3][0], colors[3][1], colors[3][2]),
-        ],
-    );
-    let idx = scratch.sky_used;
-    scratch.sky_quads[idx] = quad;
-    scratch.sky_used += 1;
-    let ptr: *mut QuadGouraud = &mut scratch.sky_quads[idx];
-    unsafe {
-        scratch
-            .ot
-            .insert(PREVIEW_SKY_SLOT, ptr.cast::<u32>(), QuadGouraud::WORDS);
-    }
-}
-
-fn push_far_vista_ring(
-    scratch: &mut PreviewScratch,
-    camera: ViewportCameraState,
-    world_camera: psx_engine::WorldCamera,
-    vista: psxed_project::ResolvedFarVistaSettings,
-    textures: &EditorTextures,
-) {
-    if !vista.enabled {
-        return;
-    }
-    let [cam_x, cam_y, cam_z] = camera.position_i32();
-    let segments = vista.segments.clamp(3, FAR_VISTA_QUAD_CAP as u8);
-    let radius = vista.radius as f32;
-    let base = (vista.rotation_degrees as f32).to_radians();
-    let step = std::f32::consts::TAU / segments as f32;
-    let y0 = cam_y.saturating_add(vista.vertical_offset);
-    let y1 = y0.saturating_add(vista.height);
-    for segment in 0..segments {
-        let a0 = base + step * segment as f32;
-        let a1 = base + step * (segment as f32 + 1.0);
-        let x0 = cam_x.saturating_add((a0.sin() * radius).round() as i32);
-        let z0 = cam_z.saturating_add((a0.cos() * radius).round() as i32);
-        let x1 = cam_x.saturating_add((a1.sin() * radius).round() as i32);
-        let z1 = cam_z.saturating_add((a1.cos() * radius).round() as i32);
-        let verts = [[x0, y1, z0], [x1, y1, z1], [x1, y0, z1], [x0, y0, z0]];
-        if !preview_vertices_in_front(world_camera, &verts) {
-            continue;
-        }
-        let projected = verts.map(|vertex| gte_scene::project_vertex(world_to_view(vertex)));
-        if projected.iter().any(|point| point.sz == 0) {
-            continue;
-        }
-        if let Some((slot, texture_width, texture_height)) =
-            far_vista_texture_slot(vista, segment, segments, textures)
-        {
-            push_far_vista_textured_quad(
-                scratch,
-                projected,
-                slot,
-                texture_width,
-                texture_height,
-                vista.tint,
-            );
-        } else {
-            push_far_vista_quad(
-                scratch,
-                [
-                    (projected[0].sx, projected[0].sy),
-                    (projected[1].sx, projected[1].sy),
-                    (projected[3].sx, projected[3].sy),
-                    (projected[2].sx, projected[2].sy),
-                ],
-                vista.tint,
-            );
-        }
-    }
-}
-
-fn far_vista_texture_slot(
-    vista: psxed_project::ResolvedFarVistaSettings,
-    segment: u8,
-    segments: u8,
-    textures: &EditorTextures,
-) -> Option<(MaterialSlot, u8, u8)> {
-    let texture_id = far_vista_texture_for_segment(vista, segment, segments)?;
-    let slot = textures.slot(texture_id)?;
-    Some((slot, slot.texture_width, slot.texture_height))
-}
-
-fn far_vista_texture_for_segment(
-    vista: psxed_project::ResolvedFarVistaSettings,
-    segment: u8,
-    segments: u8,
-) -> Option<ResourceId> {
-    let assigned_panels = vista.texture_panels.iter().any(Option::is_some);
-    if !assigned_panels {
-        return vista.texture;
-    }
-    let panel_count = active_far_vista_panel_count(&vista.texture_panels, segments);
-    if panel_count == 0 {
-        return None;
-    }
-    let panel_index = ((segment as usize) * panel_count / (segments as usize).max(1))
-        .min(panel_count.saturating_sub(1));
-    vista.texture_panels[panel_index]
-}
-
-fn active_far_vista_panel_count(
-    texture_panels: &[Option<ResourceId>; psxed_project::FAR_VISTA_TEXTURE_PANEL_COUNT],
-    segments: u8,
-) -> usize {
-    texture_panels
-        .iter()
-        .rposition(Option::is_some)
-        .map(|index| index + 1)
-        .unwrap_or(0)
-        .min(segments as usize)
-        .min(psxed_project::FAR_VISTA_TEXTURE_PANEL_COUNT)
-}
-
-fn push_far_vista_textured_quad(
-    scratch: &mut PreviewScratch,
-    projected: [psx_gte::scene::Projected; 4],
-    slot: MaterialSlot,
-    texture_width: u8,
-    texture_height: u8,
-    tint: [u8; 3],
-) {
-    let material = preview_texture_material(slot, (tint[0], tint[1], tint[2]), BlendMode::Opaque);
-    let uvs = [
-        (0, 0),
-        (texture_width.saturating_sub(1), 0),
-        (
-            texture_width.saturating_sub(1),
-            texture_height.saturating_sub(1),
-        ),
-        (0, texture_height.saturating_sub(1)),
-    ];
-    let _ = push_textured_material_tri(
-        scratch,
-        [projected[0], projected[1], projected[2]],
-        [uvs[0], uvs[1], uvs[2]],
-        material,
-        PREVIEW_FAR_VISTA_SLOT,
-    );
-    let _ = push_textured_material_tri(
-        scratch,
-        [projected[0], projected[2], projected[3]],
-        [uvs[0], uvs[2], uvs[3]],
-        material,
-        PREVIEW_FAR_VISTA_SLOT,
-    );
-}
-
-fn push_far_vista_quad(scratch: &mut PreviewScratch, verts: [(i16, i16); 4], color: [u8; 3]) {
-    if scratch.far_vista_used >= scratch.far_vista_quads.len() {
-        return;
-    }
-    let idx = scratch.far_vista_used;
-    scratch.far_vista_quads[idx] = QuadFlat::new(verts, color[0], color[1], color[2]);
-    scratch.far_vista_used += 1;
-    let ptr: *mut QuadFlat = &mut scratch.far_vista_quads[idx];
-    unsafe {
-        scratch
-            .ot
-            .insert(PREVIEW_FAR_VISTA_SLOT, ptr.cast::<u32>(), QuadFlat::WORDS);
-    }
-}
-
-/// Per-face emit: routes to the flat or textured pool based on
-/// `shade`, packing UVs only when textured.
-fn emit_face_tri(
-    scratch: &mut PreviewScratch,
-    p: [psx_gte::scene::Projected; 3],
-    uvs: [(u8, u8); 3],
-    shade: FaceShade,
-) -> bool {
-    if !face_side_visible(shade.sidedness(), p) {
-        return false;
-    }
-    match shade {
-        FaceShade::Flat { rgb, .. } => push_tri(scratch, p, rgb),
-        FaceShade::Textured {
-            slot,
-            tint,
-            blend_mode,
-            ..
-        } => push_tex_tri(scratch, p, uvs, slot, tint, blend_mode),
-    }
-}
-
-fn face_side_visible(
-    sidedness: psxed_project::MaterialFaceSidedness,
-    p: [psx_gte::scene::Projected; 3],
-) -> bool {
-    let area = projected_area(p);
-    match sidedness {
-        psxed_project::MaterialFaceSidedness::Front => area > 0,
-        psxed_project::MaterialFaceSidedness::Back => area < 0,
-        psxed_project::MaterialFaceSidedness::Both => area != 0,
-    }
-}
-
-fn projected_area(p: [psx_gte::scene::Projected; 3]) -> i32 {
-    let ax = (p[1].sx as i32) - (p[0].sx as i32);
-    let ay = (p[1].sy as i32) - (p[0].sy as i32);
-    let bx = (p[2].sx as i32) - (p[0].sx as i32);
-    let by = (p[2].sy as i32) - (p[0].sy as i32);
-    ax * by - ay * bx
-}
-
-fn preview_projected_triangle_hw_safe(p: [psx_gte::scene::Projected; 3]) -> bool {
-    let min_x = p[0].sx.min(p[1].sx).min(p[2].sx);
-    let max_x = p[0].sx.max(p[1].sx).max(p[2].sx);
-    let min_y = p[0].sy.min(p[1].sy).min(p[2].sy);
-    let max_y = p[0].sy.max(p[1].sy).max(p[2].sy);
-    min_x >= PSX_VERTEX_MIN
-        && max_x <= PSX_VERTEX_MAX
-        && min_y >= PSX_VERTEX_MIN
-        && max_y <= PSX_VERTEX_MAX
-        && ((max_x as i32) - (min_x as i32)) <= PSX_TRI_MAX_DX
-        && ((max_y as i32) - (min_y as i32)) <= PSX_TRI_MAX_DY
-}
-
-fn clamp_preview_projected_triangle(
-    p: [psx_gte::scene::Projected; 3],
-) -> [psx_gte::scene::Projected; 3] {
-    [
-        clamp_preview_projected(p[0]),
-        clamp_preview_projected(p[1]),
-        clamp_preview_projected(p[2]),
-    ]
-}
-
-fn clamp_preview_projected(p: psx_gte::scene::Projected) -> psx_gte::scene::Projected {
-    psx_gte::scene::Projected {
-        sx: p.sx.clamp(PSX_VERTEX_MIN, PSX_VERTEX_MAX),
-        sy: p.sy.clamp(PSX_VERTEX_MIN, PSX_VERTEX_MAX),
-        sz: p.sz,
-    }
-}
-
-fn split_preview_projected_triangle(
-    p: [psx_gte::scene::Projected; 3],
-    uvs: [(u8, u8); 3],
-) -> (
-    [psx_gte::scene::Projected; 3],
-    [(u8, u8); 3],
-    [psx_gte::scene::Projected; 3],
-    [(u8, u8); 3],
-) {
-    match largest_preview_projected_edge(p) {
-        0 => {
-            let mid = midpoint_projected(p[0], p[1]);
-            let uv = midpoint_uv(uvs[0], uvs[1]);
-            (
-                [p[0], mid, p[2]],
-                [uvs[0], uv, uvs[2]],
-                [mid, p[1], p[2]],
-                [uv, uvs[1], uvs[2]],
-            )
-        }
-        1 => {
-            let mid = midpoint_projected(p[1], p[2]);
-            let uv = midpoint_uv(uvs[1], uvs[2]);
-            (
-                [p[0], p[1], mid],
-                [uvs[0], uvs[1], uv],
-                [p[0], mid, p[2]],
-                [uvs[0], uv, uvs[2]],
-            )
-        }
-        _ => {
-            let mid = midpoint_projected(p[2], p[0]);
-            let uv = midpoint_uv(uvs[2], uvs[0]);
-            (
-                [p[0], p[1], mid],
-                [uvs[0], uvs[1], uv],
-                [mid, p[1], p[2]],
-                [uv, uvs[1], uvs[2]],
-            )
-        }
-    }
-}
-
-fn largest_preview_projected_edge(p: [psx_gte::scene::Projected; 3]) -> usize {
-    let mut edge = 0;
-    let mut score = preview_edge_split_score(p[0], p[1]);
-    let score_1 = preview_edge_split_score(p[1], p[2]);
-    if score_1 > score {
-        edge = 1;
-        score = score_1;
-    }
-    let score_2 = preview_edge_split_score(p[2], p[0]);
-    if score_2 > score {
-        edge = 2;
-    }
-    edge
-}
-
-fn preview_edge_split_score(a: psx_gte::scene::Projected, b: psx_gte::scene::Projected) -> i32 {
-    let dx = ((a.sx as i32) - (b.sx as i32)).abs();
-    let dy = ((a.sy as i32) - (b.sy as i32)).abs();
-    dx.max(dy.saturating_mul(2))
-}
-
-fn midpoint_projected(
-    a: psx_gte::scene::Projected,
-    b: psx_gte::scene::Projected,
-) -> psx_gte::scene::Projected {
-    psx_gte::scene::Projected {
-        sx: midpoint_i16(a.sx, b.sx),
-        sy: midpoint_i16(a.sy, b.sy),
-        sz: (((a.sz as u32) + (b.sz as u32)) / 2) as u16,
-    }
-}
-
-fn midpoint_i16(a: i16, b: i16) -> i16 {
-    (a as i32 + ((b as i32) - (a as i32)) / 2) as i16
-}
-
-fn midpoint_uv(a: (u8, u8), b: (u8, u8)) -> (u8, u8) {
-    (
-        (((a.0 as u16) + (b.0 as u16)) / 2) as u8,
-        (((a.1 as u16) + (b.1 as u16)) / 2) as u8,
-    )
-}
-
-/// Compose a [`TriTextured`] sampling `slot`'s tpage / CLUT, stash
-/// it in the static `tex_tris` arena, and chain it into the OT.
-///
-/// `tint` modulates the texel: PSX hardware computes
-/// `output = texel * tint / 0x80`, so `(0x80, 0x80, 0x80)` is a
-/// pass-through and `(0xFF, 0x60, 0x40)` saturates a grey texel
-/// toward terracotta. Textured preview uses the authored material
-/// tint so it matches the cooked runtime path; flat fallback still
-/// uses material-name colours to keep untextured faces readable.
-fn push_tex_tri(
-    scratch: &mut PreviewScratch,
-    p: [psx_gte::scene::Projected; 3],
-    uvs: [(u8, u8); 3],
-    slot: MaterialSlot,
-    tint: (u8, u8, u8),
-    blend_mode: BlendMode,
-) -> bool {
-    let avg_sz = projected_avg_sz(p);
-    push_textured_material_tri(
-        scratch,
-        p,
-        uvs,
-        preview_texture_material(slot, tint, blend_mode),
-        room_depth_slot(avg_sz),
-    )
-}
-
-fn preview_texture_material(
-    slot: MaterialSlot,
-    tint: (u8, u8, u8),
-    blend_mode: BlendMode,
-) -> TextureMaterial {
-    let material = if blend_mode.is_translucent() {
-        TextureMaterial::blended(slot.clut_word, slot.tpage_word, tint, blend_mode)
-    } else {
-        TextureMaterial::opaque(slot.clut_word, slot.tpage_word, tint)
-    };
-    material.with_texture_window(slot.texture_window)
-}
-
-fn push_shadow_tex_tri(
-    scratch: &mut PreviewScratch,
-    p: [psx_gte::scene::Projected; 3],
-    uvs: [(u8, u8); 3],
-    slot: MaterialSlot,
-) -> bool {
-    push_textured_material_tri(
-        scratch,
-        p,
-        uvs,
-        TextureMaterial::blended(
-            slot.clut_word,
-            slot.tpage_word,
-            (0x80, 0x80, 0x80),
-            BlendMode::Average,
-        )
-        .with_raw_texture(true),
-        shadow_depth_slot(projected_avg_sz(p)),
-    )
-}
-
-fn push_textured_material_tri(
-    scratch: &mut PreviewScratch,
-    p: [psx_gte::scene::Projected; 3],
-    uvs: [(u8, u8); 3],
-    material: TextureMaterial,
-    slot_idx: usize,
-) -> bool {
-    push_textured_material_tri_split(scratch, p, uvs, material, slot_idx, 0)
-}
-
-fn push_textured_material_tri_split(
-    scratch: &mut PreviewScratch,
-    p: [psx_gte::scene::Projected; 3],
-    uvs: [(u8, u8); 3],
-    material: TextureMaterial,
-    slot_idx: usize,
-    depth: u8,
-) -> bool {
-    let p = clamp_preview_projected_triangle(p);
-    if !preview_projected_triangle_hw_safe(p) {
-        if depth >= MAX_PREVIEW_HW_SPLIT_DEPTH {
-            return false;
-        }
-        let (a, a_uvs, b, b_uvs) = split_preview_projected_triangle(p, uvs);
-        let left =
-            push_textured_material_tri_split(scratch, a, a_uvs, material, slot_idx, depth + 1);
-        let right =
-            push_textured_material_tri_split(scratch, b, b_uvs, material, slot_idx, depth + 1);
-        return left || right;
-    }
-    if scratch.tex_used >= TRI_CAP {
-        return false;
-    }
-    let idx = scratch.tex_used;
-    scratch.tex_tris[idx] = TriTextured::with_material(
-        [(p[0].sx, p[0].sy), (p[1].sx, p[1].sy), (p[2].sx, p[2].sy)],
-        uvs,
-        material,
-    );
-    let packet_ptr: *mut TriTextured = &mut scratch.tex_tris[idx];
-    unsafe {
-        scratch
-            .ot
-            .insert(slot_idx, packet_ptr.cast::<u32>(), TriTextured::WORDS);
-    }
-    scratch.tex_used = idx + 1;
-    true
-}
-
-/// Compose a [`TriFlat`] from three projected vertices, store it in
-/// the next slot of the static `tris` array, and link it into the
-/// OT keyed on average screen-space depth.
-fn push_tri(
-    scratch: &mut PreviewScratch,
-    p: [psx_gte::scene::Projected; 3],
-    rgb: (u8, u8, u8),
-) -> bool {
-    push_tri_split(scratch, p, rgb, 0)
-}
-
-fn push_tri_split(
-    scratch: &mut PreviewScratch,
-    p: [psx_gte::scene::Projected; 3],
-    rgb: (u8, u8, u8),
-    depth: u8,
-) -> bool {
-    let p = clamp_preview_projected_triangle(p);
-    if !preview_projected_triangle_hw_safe(p) {
-        if depth >= MAX_PREVIEW_HW_SPLIT_DEPTH {
-            return false;
-        }
-        let (a, _, b, _) = split_preview_projected_triangle(p, [(0, 0); 3]);
-        let left = push_tri_split(scratch, a, rgb, depth + 1);
-        let right = push_tri_split(scratch, b, rgb, depth + 1);
-        return left || right;
-    }
-    if scratch.used >= TRI_CAP {
-        return false;
-    }
-    let idx = scratch.used;
-    scratch.tris[idx] = TriFlat::new(
-        [(p[0].sx, p[0].sy), (p[1].sx, p[1].sy), (p[2].sx, p[2].sy)],
-        rgb.0,
-        rgb.1,
-        rgb.2,
-    );
-    scratch.used = idx + 1;
-    // Map sz (Q0, range up to ~32K for our scenes) into the OT
-    // depth band. Smaller sz = closer = drawn last, so map to a
-    // lower OT slot index. We reserve slot OT_DEPTH-1 for the
-    // per-frame fill-rect clear and slot 0 for the hover overlay
-    // (drawn last so it tops everything), so geometry rides the
-    // 1..OT_DEPTH-1 band exclusively.
-    let slot = room_depth_slot(projected_avg_sz(p));
-    let packet_ptr: *mut TriFlat = &mut scratch.tris[idx];
-    unsafe {
-        scratch
-            .ot
-            .insert(slot, packet_ptr.cast::<u32>(), TriFlat::WORDS);
-    }
-    true
-}
-
-fn projected_avg_sz(p: [psx_gte::scene::Projected; 3]) -> u32 {
-    (p[0].sz as u32 + p[1].sz as u32 + p[2].sz as u32) / 3
-}
-
-fn room_depth_slot(avg_sz: u32) -> usize {
-    preview_geometry_depth_slot(avg_sz)
-}
-
-fn shadow_depth_slot(avg_sz: u32) -> usize {
-    preview_geometry_depth_slot(avg_sz.saturating_sub(PREVIEW_SHADOW_DEPTH_BIAS))
-}
-
-fn preview_geometry_depth_slot(avg_sz: u32) -> usize {
-    ((avg_sz as usize) >> 6).clamp(PREVIEW_GEOMETRY_SLOT_MIN, PREVIEW_GEOMETRY_SLOT_MAX)
-}
 
 #[cfg(test)]
-mod tests {
-    use super::{
-        face_side_visible, floor_anchored_model_origin, horizontal_triangle_world_points,
-        light_face, material_sized_uvs, material_texture_tint, node_room_local_origin,
-        preview_lights, preview_model_reference, preview_player_reference,
-        preview_projected_triangle_hw_safe, preview_scratch, preview_shadow_radius,
-        preview_static_model_reference, preview_vertices_in_front, push_clear, push_tri,
-        push_wall_face, room_depth_slot, rotate_image_prop_local, setup_gte_for_camera,
-        shadow_depth_slot, should_draw_culled_face_outline, wall_material_sidedness_for_edge,
-        wall_side_visible, yaw_rotation_q12, FaceShade, MaterialSlot, PreviewFog, WallEdge,
-        GRID_TILE_UV, PREVIEW_FLOOR_UVS, PREVIEW_GEOMETRY_SLOT_MAX, PREVIEW_GEOMETRY_SLOT_MIN,
-        PREVIEW_SHADOW_DEPTH_BIAS, PREVIEW_SHADOW_RADIUS_MAX, PREVIEW_SHADOW_RADIUS_MIN,
-        PREVIEW_WALL_UVS,
-    };
-    use psx_engine::{PointLightSample, WorldVertex};
-    use psx_gpu::material::BlendMode;
-    use psx_gte::scene::Projected;
-    use psxed_project::portal_rooms::PortalEdge;
-    use psxed_project::{
-        Corner, GridDirection, GridSplit, GridUvTransform, GridVerticalFace, MaterialFaceSidedness,
-        MaterialResource, NodeId, NodeKind, ProjectDocument, ResourceData, WorldGrid,
-    };
-    use psxed_ui::{ViewportCameraMode, ViewportCameraState};
-
-    fn flat(r: u8, g: u8, b: u8) -> FaceShade {
-        FaceShade::Flat {
-            rgb: (r, g, b),
-            sidedness: psxed_project::MaterialFaceSidedness::Front,
-        }
-    }
-
-    fn flat_sided(r: u8, g: u8, b: u8, sidedness: MaterialFaceSidedness) -> FaceShade {
-        FaceShade::Flat {
-            rgb: (r, g, b),
-            sidedness,
-        }
-    }
-
-    fn unpack(shade: FaceShade) -> (u8, u8, u8) {
-        match shade {
-            FaceShade::Flat { rgb, .. } => rgb,
-            FaceShade::Textured { tint, .. } => tint,
-        }
-    }
-
-    fn address_window(addr: usize) -> usize {
-        addr & !(super::OT_ADDRESS_WINDOW_BYTES - 1)
-    }
-
-    fn assert_same_ot_window<T>(label: &str, ot_window: usize, value: &T) {
-        let start = value as *const T as usize;
-        let end = start + core::mem::size_of::<T>().saturating_sub(1);
-        assert_eq!(address_window(start), ot_window, "{label} start");
-        assert_eq!(address_window(end), ot_window, "{label} end");
-    }
-
-    fn headless_preview_renderer() -> Option<psx_gpu_render::HwRenderer> {
-        let instance = wgpu::Instance::new(&wgpu::InstanceDescriptor {
-            backends: wgpu::Backends::PRIMARY,
-            ..Default::default()
-        });
-        let adapter = pollster::block_on(instance.request_adapter(&wgpu::RequestAdapterOptions {
-            power_preference: wgpu::PowerPreference::HighPerformance,
-            compatible_surface: None,
-            force_fallback_adapter: false,
-        }))?;
-        let (device, queue) = pollster::block_on(adapter.request_device(
-            &wgpu::DeviceDescriptor {
-                label: Some("editor-preview-test-device"),
-                required_features: wgpu::Features::empty(),
-                required_limits: wgpu::Limits::default(),
-                memory_hints: wgpu::MemoryHints::Performance,
-            },
-            None,
-        ))
-        .ok()?;
-        Some(psx_gpu_render::HwRenderer::new_headless(device, queue))
-    }
-
-    fn count_nonblack_rgba(rgba: &[u8]) -> usize {
-        rgba.chunks_exact(4)
-            .filter(|px| px[0] != 0 || px[1] != 0 || px[2] != 0)
-            .count()
-    }
-
-    #[test]
-    fn preview_scratch_packet_arenas_share_one_ot_address_window() {
-        let scratch = preview_scratch()
-            .lock()
-            .expect("editor preview scratch mutex");
-        let ot_window = address_window(scratch.ot.submit_head() as usize);
-
-        assert_same_ot_window("ordering table", ot_window, &scratch.ot);
-        assert_same_ot_window("clear packet", ot_window, &scratch.clear_packet);
-        assert_same_ot_window("sky quads", ot_window, &scratch.sky_quads);
-        assert_same_ot_window("far vista quads", ot_window, &scratch.far_vista_quads);
-        assert_same_ot_window("flat tris", ot_window, &scratch.tris);
-        assert_same_ot_window("textured tris", ot_window, &scratch.tex_tris);
-    }
-
-    #[test]
-    fn preview_scratch_command_log_walk_terminates() {
-        let mut scratch = preview_scratch()
-            .lock()
-            .expect("editor preview scratch mutex");
-        scratch.used = 0;
-        scratch.tex_used = 0;
-        scratch.overlay_lines.clear();
-        scratch.ot.clear();
-
-        push_clear(&mut scratch, [0, 0, 0]);
-        assert!(push_tri(
-            &mut scratch,
-            [
-                Projected {
-                    sx: 0,
-                    sy: 0,
-                    sz: 256,
-                },
-                Projected {
-                    sx: 32,
-                    sy: 0,
-                    sz: 256,
-                },
-                Projected {
-                    sx: 0,
-                    sy: 32,
-                    sz: 256,
-                },
-            ],
-            (255, 0, 0),
-        ));
-
-        let log = unsafe { psx_gpu_render::build_cmd_log(&scratch.ot) };
-        assert_eq!(log.len(), 2);
-        assert_eq!(log[0].opcode, 0x02);
-        assert_eq!(log[1].opcode, 0x20);
-    }
-
-    #[test]
-    fn demo10_preview_frame_contains_draw_commands() {
-        let repo_root = std::path::PathBuf::from(env!("CARGO_MANIFEST_DIR"))
-            .join("../../..")
-            .canonicalize()
-            .expect("repo root");
-        let project_root = repo_root.join("editor/projects/demo10");
-        let project =
-            ProjectDocument::load_from_path(project_root.join("project.ron")).expect("demo10");
-        let mut textures = crate::editor_textures::EditorTextures::new();
-        textures.refresh(&project, &project_root);
-        textures.refresh_models(&project, &project_root);
-        let mut assets = crate::editor_assets::EditorAssets::new();
-        assets.refresh(&project, &project_root);
-
-        let camera = ViewportCameraState {
-            mode: ViewportCameraMode::Orbit,
-            yaw_q12: 320,
-            pitch_q12: 300,
-            radius: 8192,
-            target: [2048, 512, 2048],
-            position: [0, 0, 0],
-        };
-        let empty_hidden = std::collections::HashSet::new();
-        let frame = super::build_phase1_frame(
-            &project,
-            camera,
-            true,
-            true,
-            true,
-            true,
-            true,
-            true,
-            &empty_hidden,
-            None,
-            0,
-            NodeId::ROOT,
-            None,
-            None,
-            &[],
-            &[],
-            None,
-            &[],
-            None,
-            &[],
-            None,
-            &textures,
-            &assets,
-        );
-
-        let draw_count = frame
-            .cmd_log
-            .iter()
-            .filter(|entry| matches!(entry.opcode, 0x20..=0x7F))
-            .count();
-        let mut translator = psx_gpu_render::Translator::new();
-        let translated = translator.translate(&frame.cmd_log);
-        let nonblack_vertices = translated
-            .vertices
-            .iter()
-            .filter(|v| v.color[0] != 0 || v.color[1] != 0 || v.color[2] != 0)
-            .count();
-        assert!(
-            draw_count > 0,
-            "demo10 preview should emit draw commands; opcodes={:?}",
-            frame
-                .cmd_log
-                .iter()
-                .map(|entry| entry.opcode)
-                .collect::<Vec<_>>()
-        );
-        assert!(
-            translated.total() > 0,
-            "demo10 preview should translate to vertices"
-        );
-        assert!(
-            nonblack_vertices > 0,
-            "demo10 preview should contain visible non-black vertices"
-        );
-
-        if let Some(mut renderer) = headless_preview_renderer() {
-            assert!(renderer.set_internal_scale(2, None));
-            let vram = textures.vram_words();
-            renderer.render_frame(&emulator_core::Gpu::new(), &frame.cmd_log, vram);
-            let scale = renderer.internal_scale();
-            let (_, _, rgba) = renderer.read_subrect_rgba8(0, 0, 320 * scale, 240 * scale);
-            assert!(
-                count_nonblack_rgba(&rgba) > 0,
-                "demo10 preview should render non-black pixels"
-            );
-        }
-    }
-
-    fn fog(rgb: (u8, u8, u8), near: i32, far: i32) -> PreviewFog {
-        PreviewFog {
-            enabled: true,
-            rgb,
-            near,
-            far,
-        }
-    }
-
-    #[test]
-    fn image_prop_overlay_rotation_applies_roll() {
-        let rotated = rotate_image_prop_local([0, 512, 0], [0, 0, 1024]);
-        assert_eq!(rotated, [-512, 0, 0]);
-    }
-
-    #[test]
-    fn image_prop_overlay_rotation_applies_yaw_q12_quarter_turn() {
-        let rotated = rotate_image_prop_local([512, 0, 0], [0, 1024, 0]);
-        assert_eq!(rotated, [0, 0, -512]);
-    }
-
-    #[test]
-    fn model_preview_yaw_matrix_uses_q12_quarter_turn() {
-        let rotation = yaw_rotation_q12(1024);
-        assert_eq!(rotation.m, [[0, 0, 4096], [0, 4096, 0], [-4096, 0, 0]]);
-    }
-
-    fn projected(sx: i16, sy: i16) -> Projected {
-        Projected { sx, sy, sz: 100 }
-    }
-
-    #[test]
-    fn textured_preview_uses_authored_material_tint() {
-        let mut project = ProjectDocument::new("test");
-        let texture = project.add_resource(
-            "Brick Texture",
-            ResourceData::Texture {
-                psxt_path: "brick.psxt".to_string(),
-            },
-        );
-        let mut material = MaterialResource::opaque(Some(texture));
-        material.tint = [0x60, 0x70, 0x90];
-        let material = project.add_resource("Brick Wall", ResourceData::Material(material));
-
-        assert_eq!(
-            material_texture_tint(&project, material),
-            (0x60, 0x70, 0x90)
-        );
-    }
-
-    #[test]
-    fn wall_preview_uvs_use_runtime_grid_tile_span() {
-        let transform = GridUvTransform {
-            span: [0, 128],
-            ..GridUvTransform::IDENTITY
-        };
-
-        assert_eq!(
-            transform.apply_to_quad(PREVIEW_WALL_UVS),
-            [(0, 128), (64, 128), (64, 0), (0, 0)]
-        );
-    }
-
-    #[test]
-    fn horizontal_preview_points_use_triangle_local_heights() {
-        let corners = psxed_project::horizontal_triangle_corners(GridSplit::NorthWestSouthEast, 0);
-        assert_eq!(corners, [Corner::NW, Corner::NE, Corner::SE]);
-        assert_eq!(
-            horizontal_triangle_world_points([0, 1024, 0, 1024], corners, [64, 128, 192]),
-            [[0, 64, 1024], [1024, 128, 1024], [1024, 192, 0]]
-        );
-    }
-
-    #[test]
-    fn portal_seam_runs_coalesce_connected_edges() {
-        let runs = super::portal_seam_runs(vec![
-            PortalEdge {
-                x: 2,
-                z: 0,
-                direction: GridDirection::North,
-            },
-            PortalEdge {
-                x: 0,
-                z: 0,
-                direction: GridDirection::North,
-            },
-            PortalEdge {
-                x: 1,
-                z: 0,
-                direction: GridDirection::North,
-            },
-            PortalEdge {
-                x: 4,
-                z: 2,
-                direction: GridDirection::East,
-            },
-            PortalEdge {
-                x: 4,
-                z: 3,
-                direction: GridDirection::East,
-            },
-        ]);
-
-        assert_eq!(
-            runs,
-            vec![
-                super::PortalSeamRun {
-                    start: PortalEdge {
-                        x: 0,
-                        z: 0,
-                        direction: GridDirection::North,
-                    },
-                    len: 3,
-                },
-                super::PortalSeamRun {
-                    start: PortalEdge {
-                        x: 4,
-                        z: 2,
-                        direction: GridDirection::East,
-                    },
-                    len: 2,
-                },
-            ]
-        );
-    }
-
-    #[test]
-    fn portal_edge_height_span_uses_real_adjacent_geometry() {
-        let mut grid = WorldGrid::empty(2, 1, 2048);
-        grid.set_floor(0, 0, -128, None);
-        grid.set_floor(1, 0, 64, None);
-        grid.ensure_sector(1, 0)
-            .unwrap()
-            .walls
-            .east
-            .push(GridVerticalFace::flat(0, 3328, None));
-
-        let corners =
-            super::portal_edge_wall_corners_for_world_cell(&grid, 0, 0, GridDirection::East)
-                .expect("portal wall corners");
-
-        assert_eq!(corners[0][1], -104);
-        assert_eq!(corners[1][1], -104);
-        assert_eq!(corners[2][1], 3328);
-        assert_eq!(corners[3][1], 3328);
-    }
-
-    #[test]
-    fn visible_room_grids_keeps_all_non_hidden_rooms() {
-        let mut project = ProjectDocument::new("test");
-        let scene = project.active_scene_mut();
-        let room_a = scene.add_node(
-            scene.root,
-            "Room A",
-            NodeKind::Room {
-                grid: WorldGrid::empty(1, 1, 1024),
-            },
-        );
-        let room_b = scene.add_node(
-            scene.root,
-            "Room B",
-            NodeKind::Room {
-                grid: WorldGrid::empty(1, 1, 1024),
-            },
-        );
-
-        let mut hidden = std::collections::HashSet::new();
-        let visible = super::visible_room_grids(&project, &hidden)
-            .into_iter()
-            .map(|(id, _)| id)
-            .collect::<Vec<_>>();
-        assert_eq!(visible, vec![room_a, room_b]);
-
-        hidden.insert(room_a);
-        let visible = super::visible_room_grids(&project, &hidden)
-            .into_iter()
-            .map(|(id, _)| id)
-            .collect::<Vec<_>>();
-        assert_eq!(visible, vec![room_b]);
-    }
-
-    #[test]
-    fn preview_room_grids_walks_bounded_portal_neighborhood() {
-        let mut project = ProjectDocument::new("test");
-        let scene = project.active_scene_mut();
-        let room_a = scene.add_node(
-            scene.root,
-            "Room A",
-            NodeKind::Room {
-                grid: WorldGrid::empty(1, 1, 1024),
-            },
-        );
-        let room_b = scene.add_node(
-            scene.root,
-            "Room B",
-            NodeKind::Room {
-                grid: WorldGrid::empty(1, 1, 1024),
-            },
-        );
-        let room_c = scene.add_node(
-            scene.root,
-            "Room C",
-            NodeKind::Room {
-                grid: WorldGrid::empty(1, 1, 1024),
-            },
-        );
-        let room_d = scene.add_node(
-            scene.root,
-            "Room D",
-            NodeKind::Room {
-                grid: WorldGrid::empty(1, 1, 1024),
-            },
-        );
-        let room_e = scene.add_node(
-            scene.root,
-            "Room E",
-            NodeKind::Room {
-                grid: WorldGrid::empty(1, 1, 1024),
-            },
-        );
-        let room_far = scene.add_node(
-            scene.root,
-            "Room Far",
-            NodeKind::Room {
-                grid: WorldGrid::empty(1, 1, 1024),
-            },
-        );
-        for (source, target) in [
-            (room_a, room_b),
-            (room_b, room_c),
-            (room_c, room_d),
-            (room_d, room_e),
-        ] {
-            scene.add_node(
-                source,
-                "Portal",
-                NodeKind::Portal {
-                    target_room: Some(target),
-                    target_entry: String::new(),
-                    entry_name: String::new(),
-                    geometry: None,
-                },
-            );
-        }
-
-        let hidden = std::collections::HashSet::new();
-        let visible = super::preview_room_grids(
-            &project,
-            &hidden,
-            Some(room_a),
-            0,
-            NodeId::ROOT,
-            None,
-            &[],
-            &[],
-        )
-        .into_iter()
-        .map(|entry| entry.room)
-        .collect::<Vec<_>>();
-
-        assert_eq!(visible, vec![room_a, room_b, room_c, room_d]);
-        assert!(!visible.contains(&room_e));
-        assert!(!visible.contains(&room_far));
-    }
-
-    /// Sims-style floor view: the active floor is the working plane at
-    /// y_offset 0 (so the ray pick, which tests the active floor's
-    /// floor-local faces, aligns with what's drawn), floors BELOW render
-    /// descending (negative offset), and floors ABOVE are hidden. This is
-    /// what keeps selection on the floor you're editing.
-    #[test]
-    fn preview_room_grids_shows_active_floor_and_below_only() {
-        let mut project = ProjectDocument::new("test");
-        let scene = project.active_scene_mut();
-        let mut grid = WorldGrid::empty(1, 1, 1024);
-        grid.push_floor(); // floor 1
-        grid.push_floor(); // floor 2
-        let room = scene.add_node(scene.root, "Stacked", NodeKind::Room { grid });
-
-        let hidden = std::collections::HashSet::new();
-        // Active floor = 1 (middle of three).
-        let floors = super::preview_room_grids(
-            &project,
-            &hidden,
-            Some(room),
-            1,
-            NodeId::ROOT,
-            None,
-            &[],
-            &[],
-        );
-        let entries: Vec<(usize, i32, bool)> = floors
-            .iter()
-            .filter(|f| f.room == room)
-            .map(|f| (f.floor_index, f.y_offset, f.active))
-            .collect();
-
-        // Floors 0 and 1 only (the active floor and the one below); floor
-        // 2 (above) is hidden.
-        assert_eq!(entries.len(), 2, "active + below only: {entries:?}");
-        assert!(
-            entries
-                .iter()
-                .any(|&(i, off, act)| i == 1 && off == 0 && act),
-            "active floor 1 at y_offset 0: {entries:?}"
-        );
-        let below = entries
-            .iter()
-            .find(|&&(i, _, _)| i == 0)
-            .expect("floor 0 below");
-        assert!(
-            below.1 < 0,
-            "floor below descends (negative y_offset): {below:?}"
-        );
-        assert!(
-            !entries.iter().any(|&(i, _, _)| i == 2),
-            "floor above hidden"
-        );
-    }
-
-    #[test]
-    fn component_model_reference_reads_renderer_and_animator_children() {
-        let mut project = ProjectDocument::new("test");
-        let model_id = project.add_resource(
-            "Dummy",
-            ResourceData::Texture {
-                psxt_path: "dummy.psxt".to_string(),
-            },
-        );
-        let scene = project.active_scene_mut();
-        let actor = scene.add_node(scene.root, "Enemy", NodeKind::Entity);
-        let renderer = scene.add_node(
-            actor,
-            "Model Renderer",
-            NodeKind::ModelRenderer {
-                model: Some(model_id),
-                material: None,
-                visual_offset: [0; 3],
-                visual_scale_q8: psxed_project::MODEL_SCALE_ONE_Q8,
-            },
-        );
-        let animator = scene.add_node(
-            actor,
-            "Animator",
-            NodeKind::Animator {
-                clip: Some(3),
-                action_clips: Vec::new(),
-                autoplay: true,
-                pose_frame: 0,
-            },
-        );
-
-        let scene = project.active_scene();
-        let reference = preview_model_reference(scene, scene.node(actor).unwrap()).unwrap();
-
-        assert_eq!(reference.model_id, model_id);
-        assert_eq!(reference.clip_override, Some(3));
-        assert!(reference.autoplay);
-        assert_eq!(reference.renderer_node, Some(renderer));
-        assert_eq!(reference.animator_node, Some(animator));
-    }
-
-    #[test]
-    fn component_player_reference_reads_controller_renderer_and_animator_children() {
-        let mut project = ProjectDocument::new("test");
-        let character_id = project.add_resource(
-            "Dummy",
-            ResourceData::Texture {
-                psxt_path: "dummy.psxt".to_string(),
-            },
-        );
-        let scene = project.active_scene_mut();
-        let actor = scene.add_node(scene.root, "Player", NodeKind::Entity);
-        let controller = scene.add_node(
-            actor,
-            "Character Controller",
-            NodeKind::CharacterController {
-                character: Some(character_id),
-                player: true,
-                settings: Default::default(),
-            },
-        );
-        let renderer = scene.add_node(
-            actor,
-            "Model Renderer",
-            NodeKind::ModelRenderer {
-                model: None,
-                material: None,
-                visual_offset: [0; 3],
-                visual_scale_q8: psxed_project::MODEL_SCALE_ONE_Q8,
-            },
-        );
-        let animator = scene.add_node(
-            actor,
-            "Animator",
-            NodeKind::Animator {
-                clip: Some(2),
-                action_clips: Vec::new(),
-                autoplay: false,
-                pose_frame: 0,
-            },
-        );
-
-        let scene = project.active_scene();
-        let reference = preview_player_reference(scene, scene.node(actor).unwrap()).unwrap();
-
-        assert_eq!(reference.character, Some(character_id));
-        assert_eq!(reference.controller_node, Some(controller));
-        assert_eq!(reference.renderer_node, Some(renderer));
-        assert_eq!(reference.animator_node, Some(animator));
-        assert_eq!(reference.clip_override, Some(2));
-        assert!(!reference.autoplay);
-    }
-
-    #[test]
-    fn player_controlled_entity_does_not_static_preview_model_renderer() {
-        let mut project = ProjectDocument::new("test");
-        let model_id = project.add_resource(
-            "Dummy Model",
-            ResourceData::Texture {
-                psxt_path: "dummy.psxt".to_string(),
-            },
-        );
-        let character_id = project.add_resource(
-            "Dummy Character",
-            ResourceData::Texture {
-                psxt_path: "dummy-character.psxt".to_string(),
-            },
-        );
-        let scene = project.active_scene_mut();
-        let actor = scene.add_node(scene.root, "Player", NodeKind::Entity);
-        scene.add_node(
-            actor,
-            "Model Renderer",
-            NodeKind::ModelRenderer {
-                model: Some(model_id),
-                material: None,
-                visual_offset: [0; 3],
-                visual_scale_q8: psxed_project::MODEL_SCALE_ONE_Q8,
-            },
-        );
-        scene.add_node(
-            actor,
-            "Character Controller",
-            NodeKind::CharacterController {
-                character: Some(character_id),
-                player: true,
-                settings: Default::default(),
-            },
-        );
-
-        let scene = project.active_scene();
-        let actor_node = scene.node(actor).unwrap();
-        assert!(
-            preview_model_reference(scene, actor_node).is_some(),
-            "the raw renderer reference is still present"
-        );
-        assert!(
-            preview_static_model_reference(scene, actor_node).is_none(),
-            "player-controlled renderers are drawn by the player preview path"
-        );
-    }
-
-    #[test]
-    fn point_light_uses_own_transform() {
-        let mut project = ProjectDocument::new("test");
-        let scene = project.active_scene_mut();
-        let host = scene.add_node(scene.root, "Lamp", NodeKind::Entity);
-        scene.node_mut(host).unwrap().transform.translation = [2.0, 0.5, 3.0];
-        let light = scene.add_node(
-            scene.root,
-            "Point Light",
-            NodeKind::PointLight {
-                color: [1, 2, 3],
-                intensity: 0.75,
-                radius: 4.0,
-            },
-        );
-        scene.node_mut(light).unwrap().transform.translation = [99.0, 99.0, 99.0];
-
-        let hidden = std::collections::HashSet::new();
-        let lights = preview_lights(project.active_scene(), &hidden);
-
-        assert_eq!(lights.len(), 1);
-        assert_eq!(lights[0].host_id, light);
-        assert_eq!(lights[0].transform.translation, [99.0, 99.0, 99.0]);
-        assert_eq!(lights[0].color, [1, 2, 3]);
-        assert_eq!(lights[0].intensity, 0.75);
-        assert_eq!(lights[0].radius, 4.0);
-    }
-
-    #[test]
-    fn face_sidedness_matches_runtime_winding_convention() {
-        let front = [projected(0, 0), projected(10, 0), projected(0, 10)];
-        let back = [front[0], front[2], front[1]];
-
-        assert!(face_side_visible(MaterialFaceSidedness::Front, front));
-        assert!(!face_side_visible(MaterialFaceSidedness::Front, back));
-        assert!(!face_side_visible(MaterialFaceSidedness::Back, front));
-        assert!(face_side_visible(MaterialFaceSidedness::Back, back));
-        assert!(face_side_visible(MaterialFaceSidedness::Both, front));
-        assert!(face_side_visible(MaterialFaceSidedness::Both, back));
-    }
-
-    #[test]
-    fn preview_rejects_triangles_outside_psx_packet_extent() {
-        assert!(preview_projected_triangle_hw_safe([
-            projected(0, 0),
-            projected(64, 0),
-            projected(0, 64),
-        ]));
-        assert!(!preview_projected_triangle_hw_safe([
-            projected(-1024, 0),
-            projected(1023, 0),
-            projected(0, 16),
-        ]));
-        assert!(!preview_projected_triangle_hw_safe([
-            projected(0, -512),
-            projected(16, 512),
-            projected(0, 0),
-        ]));
-    }
-
-    #[test]
-    fn editor_cardinal_wall_front_material_renders_from_owning_cell() {
-        let cases = [
-            (WallEdge::North, [512, 512, 512], 2048, [512, 512, 1536], 0),
-            (
-                WallEdge::East,
-                [512, 512, 512],
-                3072,
-                [1536, 512, 512],
-                1024,
-            ),
-            (WallEdge::South, [512, 512, 512], 0, [512, 512, -512], 2048),
-            (
-                WallEdge::West,
-                [512, 512, 512],
-                1024,
-                [-512, 512, 512],
-                3072,
-            ),
-        ];
-
-        for (edge, inside_pos, inside_yaw, outside_pos, outside_yaw) in cases {
-            assert!(
-                wall_face_emits_from_camera(
-                    edge,
-                    inside_pos,
-                    inside_yaw,
-                    MaterialFaceSidedness::Front
-                ),
-                "{edge:?} wall front material should render from inside the owning cell"
-            );
-            assert!(
-                !wall_face_emits_from_camera(
-                    edge,
-                    inside_pos,
-                    inside_yaw,
-                    MaterialFaceSidedness::Back
-                ),
-                "{edge:?} wall back material should not render from inside the owning cell"
-            );
-            assert!(
-                !wall_face_emits_from_camera(
-                    edge,
-                    outside_pos,
-                    outside_yaw,
-                    MaterialFaceSidedness::Front
-                ),
-                "{edge:?} wall front material should not render from outside the owning cell"
-            );
-            assert!(
-                wall_face_emits_from_camera(
-                    edge,
-                    outside_pos,
-                    outside_yaw,
-                    MaterialFaceSidedness::Back
-                ),
-                "{edge:?} wall back material should render from outside the owning cell"
-            );
-        }
-    }
-
-    #[test]
-    fn editor_diagonal_wall_materials_are_forced_double_sided() {
-        for edge in [WallEdge::NorthWestSouthEast, WallEdge::NorthEastSouthWest] {
-            assert_eq!(
-                wall_material_sidedness_for_edge(MaterialFaceSidedness::Front, edge),
-                MaterialFaceSidedness::Both
-            );
-            assert_eq!(
-                wall_material_sidedness_for_edge(MaterialFaceSidedness::Back, edge),
-                MaterialFaceSidedness::Both
-            );
-            assert!(wall_side_visible(
-                MaterialFaceSidedness::Front,
-                [0, 1024, 0, 1024],
-                edge,
-                [256, 512, 256]
-            ));
-            assert!(wall_side_visible(
-                MaterialFaceSidedness::Back,
-                [0, 1024, 0, 1024],
-                edge,
-                [768, 512, 768]
-            ));
-        }
-    }
-
-    fn wall_face_emits_from_camera(
-        edge: WallEdge,
-        position: [i32; 3],
-        yaw_q12: u16,
-        sidedness: MaterialFaceSidedness,
-    ) -> bool {
-        let camera = setup_gte_for_camera(ViewportCameraState {
-            mode: ViewportCameraMode::Free,
-            yaw_q12,
-            pitch_q12: 0,
-            radius: 1024,
-            target: [512, 512, 512],
-            position,
-        });
-        let mut scratch = preview_scratch()
-            .lock()
-            .expect("editor preview scratch mutex");
-        scratch.used = 0;
-        scratch.tex_used = 0;
-        scratch.overlay_lines.clear();
-        scratch.ot.clear();
-
-        push_wall_face(
-            &mut scratch,
-            camera,
-            [0, 1024, 0, 1024],
-            edge,
-            [384, 384, 640, 640],
-            None,
-            GridUvTransform::default(),
-            flat_sided(128, 128, 128, sidedness),
-            position,
-        );
-
-        scratch.used > 0 || scratch.tex_used > 0
-    }
-
-    #[test]
-    fn preview_near_guard_rejects_vertices_behind_camera() {
-        let camera = setup_gte_for_camera(ViewportCameraState {
-            mode: ViewportCameraMode::Free,
-            yaw_q12: 0,
-            pitch_q12: 0,
-            radius: 1024,
-            target: [0, 0, 0],
-            position: [0, 0, 0],
-        });
-
-        assert!(preview_vertices_in_front(
-            camera,
-            &[[0, 0, -64], [16, 0, -64]]
-        ));
-        assert!(!preview_vertices_in_front(
-            camera,
-            &[[0, 0, -64], [0, 0, 16]]
-        ));
-    }
-
-    #[test]
-    fn culled_room_face_outline_respects_preview_toggle() {
-        assert!(should_draw_culled_face_outline(
-            true,
-            flat_sided(128, 128, 128, MaterialFaceSidedness::Front)
-        ));
-        assert!(should_draw_culled_face_outline(
-            true,
-            flat_sided(128, 128, 128, MaterialFaceSidedness::Back)
-        ));
-        assert!(!should_draw_culled_face_outline(
-            false,
-            flat_sided(128, 128, 128, MaterialFaceSidedness::Back)
-        ));
-        assert!(!should_draw_culled_face_outline(
-            true,
-            flat_sided(128, 128, 128, MaterialFaceSidedness::Both)
-        ));
-    }
-
-    #[test]
-    fn preview_depth_slots_share_world_geometry_band() {
-        assert_eq!(room_depth_slot(0), PREVIEW_GEOMETRY_SLOT_MIN);
-        assert_eq!(room_depth_slot(u32::MAX), PREVIEW_GEOMETRY_SLOT_MAX);
-        assert!(shadow_depth_slot(2048) < room_depth_slot(2048));
-        assert_eq!(shadow_depth_slot(0), PREVIEW_GEOMETRY_SLOT_MIN);
-        assert_eq!(PREVIEW_SHADOW_DEPTH_BIAS, 128);
-    }
-
-    #[test]
-    fn preview_shadow_radius_matches_runtime_scale() {
-        assert_eq!(preview_shadow_radius(1), PREVIEW_SHADOW_RADIUS_MIN);
-        assert_eq!(preview_shadow_radius(2048), PREVIEW_SHADOW_RADIUS_MAX);
-        assert_eq!(preview_shadow_radius(200), 250);
-    }
-
-    #[test]
-    fn node_room_local_origin_matches_origin_aware_grid_conversion() {
-        let mut grid = WorldGrid::stone_room(4, 7, 1024, None, None);
-        grid.origin = [-1, -3];
-        let translation = [1.0, 0.25, 0.85];
-        let transform = psxed_project::Transform3 {
-            translation,
-            ..psxed_project::Transform3::default()
-        };
-
-        let origin = node_room_local_origin(&grid, &transform);
-        let expected = grid.editor_to_room_local([translation[0], translation[2]]);
-
-        assert_eq!(origin.x, expected[0] as i32);
-        assert_eq!(origin.y, 256);
-        assert_eq!(origin.z, expected[2] as i32);
-        assert_ne!(
-            (origin.x, origin.z),
-            (
-                ((translation[0] + grid.width as f32 * 0.5) * grid.sector_size as f32) as i32,
-                ((translation[2] + grid.depth as f32 * 0.5) * grid.sector_size as f32) as i32,
-            ),
-            "regression guard: old half-grid-only conversion ignores grid.origin"
-        );
-    }
-
-    #[test]
-    fn floor_anchored_model_origin_offsets_by_half_world_height() {
-        let origin = floor_anchored_model_origin(WorldVertex::new(10, 0, 20), 1024);
-        assert_eq!(origin, WorldVertex::new(10, 512, 20));
-    }
-
-    #[test]
-    fn floor_anchored_model_origin_ignores_negative_height() {
-        let origin = floor_anchored_model_origin(WorldVertex::new(10, 32, 20), -128);
-        assert_eq!(origin, WorldVertex::new(10, 32, 20));
-    }
-
-    #[test]
-    fn preview_fog_blends_after_near_plane() {
-        let fog = fog((10, 20, 30), 100, 300);
-
-        assert_eq!(fog.apply_rgb((110, 120, 130), 100), (110, 120, 130));
-        assert_eq!(fog.apply_rgb((110, 120, 130), 200), (60, 70, 80));
-        assert_eq!(fog.apply_rgb((110, 120, 130), 300), (10, 20, 30));
-        assert_eq!(fog.apply_rgb((110, 120, 130), 900), (10, 20, 30));
-    }
-
-    #[test]
-    fn preview_fog_applies_to_flat_and_textured_tints() {
-        let fog = fog((0, 0, 0), 0, 256);
-        let flat = fog.apply_shade(flat(128, 64, 32), 128);
-        let textured = fog.apply_shade(
-            FaceShade::Textured {
-                slot: MaterialSlot {
-                    tpage_word: 0,
-                    clut_word: 0,
-                    texture_window: psx_gpu::material::TextureWindow::NONE,
-                    texture_width: 64,
-                    texture_height: 64,
-                },
-                tint: (128, 64, 32),
-                blend_mode: BlendMode::Opaque,
-                sidedness: psxed_project::MaterialFaceSidedness::Front,
-            },
-            128,
-        );
-
-        assert_eq!(unpack(flat), (64, 32, 16));
-        assert_eq!(unpack(textured), (64, 32, 16));
-    }
-
-    #[test]
-    fn material_sized_uvs_stretch_32px_texture_once_by_default() {
-        let shade = FaceShade::Textured {
-            slot: MaterialSlot {
-                tpage_word: 0,
-                clut_word: 0,
-                texture_window: psx_gpu::material::TextureWindow::NONE,
-                texture_width: 32,
-                texture_height: 32,
-            },
-            tint: (128, 128, 128),
-            blend_mode: BlendMode::Opaque,
-            sidedness: psxed_project::MaterialFaceSidedness::Both,
-        };
-        assert_eq!(
-            material_sized_uvs(shade, PREVIEW_FLOOR_UVS),
-            [(0, 0), (32, 0), (32, 32), (0, 32)]
-        );
-    }
-
-    #[test]
-    fn material_sized_uvs_preserve_authored_repeat_count() {
-        let shade = FaceShade::Textured {
-            slot: MaterialSlot {
-                tpage_word: 0,
-                clut_word: 0,
-                texture_window: psx_gpu::material::TextureWindow::NONE,
-                texture_width: 32,
-                texture_height: 64,
-            },
-            tint: (128, 128, 128),
-            blend_mode: BlendMode::Opaque,
-            sidedness: psxed_project::MaterialFaceSidedness::Both,
-        };
-        assert_eq!(
-            material_sized_uvs(
-                shade,
-                [(0, 0), (128, 0), (128, GRID_TILE_UV), (0, GRID_TILE_UV)]
-            ),
-            [(0, 0), (64, 0), (64, GRID_TILE_UV), (0, GRID_TILE_UV)]
-        );
-    }
-
-    #[test]
-    fn light_face_no_lights_ambient_32_is_not_white() {
-        // Regression: pre-fix the `ambient * 256` bug saturated
-        // every face to 255. With the new convention an unlit
-        // face at ambient 32 should render at ~32, not white.
-        let base = flat(128, 128, 128);
-        let lit = light_face(base, [0, 0, 0], &[], [32, 32, 32]);
-        let (r, g, b) = unpack(lit);
-        assert!(r < 64 && g < 64 && b < 64, "got ({r}, {g}, {b})");
-    }
-
-    #[test]
-    fn light_face_ambient_128_is_neutral() {
-        // 128 ambient is the neutral PSX-tint value; an unlit
-        // 128-base material should land back at exactly 128.
-        let lit = light_face(flat(128, 128, 128), [0, 0, 0], &[], [128, 128, 128]);
-        assert_eq!(unpack(lit), (128, 128, 128));
-    }
-
-    #[test]
-    fn light_face_zero_ambient_zero_lights_black() {
-        let lit = light_face(flat(255, 255, 255), [0, 0, 0], &[], [0, 0, 0]);
-        assert_eq!(unpack(lit), (0, 0, 0));
-    }
-
-    #[test]
-    fn light_face_point_light_inside_radius_brightens() {
-        // White light at the face centre with neutral base
-        // should land at saturating-bright since contribution
-        // (255 × 256 × 256) / 65536 = 255 dominates ambient.
-        let light = PointLightSample::from_color_intensity_q8([0, 0, 0], 100, [255, 255, 255], 256);
-        let lit = light_face(flat(128, 128, 128), [0, 0, 0], &[light], [32, 32, 32]);
-        let (r, g, b) = unpack(lit);
-        assert!(r > 200 && g > 200 && b > 200, "got ({r}, {g}, {b})");
-    }
-
-    #[test]
-    fn light_face_point_light_outside_radius_zero() {
-        // Place the face well outside the light's radius; the
-        // contribution must be exactly zero. Output should
-        // match the no-lights case.
-        let light = PointLightSample::from_color_intensity_q8([0, 0, 0], 100, [255, 255, 255], 256);
-        let lit = light_face(flat(128, 128, 128), [10000, 0, 0], &[light], [32, 32, 32]);
-        let baseline = light_face(flat(128, 128, 128), [10000, 0, 0], &[], [32, 32, 32]);
-        assert_eq!(unpack(lit), unpack(baseline));
-    }
-
-    #[test]
-    fn light_face_two_lights_accumulate_and_clamp() {
-        let l = PointLightSample::from_color_intensity_q8([0, 0, 0], 100, [255, 255, 255], 256);
-        let lit = light_face(flat(255, 255, 255), [0, 0, 0], &[l, l], [128, 128, 128]);
-        let (r, g, b) = unpack(lit);
-        // Even with two saturating lights, output never
-        // exceeds 255 per channel.
-        assert_eq!((r, g, b), (255, 255, 255));
-    }
-}
+mod tests;
